requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
whitenoise>=6.7,<7.0
gunicorn>=21.2,<22.0
psycopg[binary,pool]>=3.2,<4.0
dj-database-url>=2.2,<3.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
